refactor(suite): create @suite-common/fee-types and move fee types out of wallet-types
