+++ packages/suite/src/components/earn/staking/tron/hooks/useTronStakeFees.ts
@@ -1,11 +1,12 @@
 import { useEffect, useMemo, useState } from 'react';
 
+import { type FeeInfo } from '@suite-common/fee-types';
 import {
     composeTronFreezeFeeLevelsThunk,
     composeTronVoteFeeLevelsThunk,
     selectRawNetworkFeeInfo,
 } from '@suite-common/wallet-core';
-import { type FeeInfo, type PrecomposedLevels } from '@suite-common/wallet-types';
+import { type PrecomposedLevels } from '@suite-common/wallet-types';
 import { getConvertedOrDefaultFeeInfo } from '@suite-common/wallet-utils';
 import { exhaustive } from '@trezor/type-utils';
