+++ packages/suite/src/components/wallet/Fees/CollapsibleFees/CustomFee/CurrentFee.tsx
@@ -1,6 +1,6 @@
 import { Translation } from '@suite/intl';
 import { useFeesContext } from '@suite-common/fee';
-import { type FeeInfo } from '@suite-common/wallet-types';
+import { type FeeInfo } from '@suite-common/fee-types';
 import { isEip1559 } from '@suite-common/wallet-utils';
 import { Icon, Row, Text } from '@trezor/components';
 import { FeeRate } from '@trezor/product-components';
