+++ packages/suite/src/views/wallet/trading/common/TradingDetail/utils.ts
@@ -1,6 +1,6 @@
+import { type FeeInfo } from '@suite-common/fee-types';
 import { type TradingComposedTransactionInfo } from '@suite-common/trading';
 import { type NetworkType } from '@suite-common/wallet-config';
-import { type FeeInfo } from '@suite-common/wallet-types';
 
 export const getTxEstimatedTimeSeconds = (
     networkType: NetworkType | undefined,
