+++ packages/suite/tsconfig.json
@@ -57,6 +57,9 @@
             "path": "../../suite-common/earn-staking-api"
         },
         { "path": "../../suite-common/fee" },
+        {
+            "path": "../../suite-common/fee-types"
+        },
         { "path": "../../suite-common/feedback" },
         {
             "path": "../../suite-common/fiat-services"
