+++ suite-common/fee-types/src/index.ts
@@ -1,0 +1,1 @@
+export type * from './feeTypes';
