+++ suite-common/fee-types/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": { "outDir": "libDev" },
+    "references": [
+        { "path": "../wallet-config" },
+        { "path": "../../packages/connect" }
+    ]
+}
