+++ suite-common/test-utils/package.json
@@ -15,6 +15,7 @@
         "@reduxjs/toolkit": "2.11.2",
         "@suite-common/analytics": "workspace:*",
         "@suite-common/bip329-types": "workspace:*",
+        "@suite-common/fee-types": "workspace:*",
         "@suite-common/platform-encryption": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/suite-sync-types": "workspace:*",
