+++ suite-common/test-utils/tsconfig.json
@@ -4,6 +4,7 @@
     "references": [
         { "path": "../analytics" },
         { "path": "../bip329-types" },
+        { "path": "../fee-types" },
         { "path": "../platform-encryption" },
         { "path": "../redux-utils" },
         { "path": "../suite-sync-types" },
