+++ suite-common/trading/package.json
@@ -15,6 +15,7 @@
     "dependencies": {
         "@reduxjs/toolkit": "2.11.2",
         "@suite-common/device": "workspace:*",
+        "@suite-common/fee-types": "workspace:*",
         "@suite-common/geolocation": "workspace:*",
         "@suite-common/react-query": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
