+++ suite-common/trading/src/thunks/common/__tests__/recomposeAndSignTxThunk.test.ts
@@ -1,12 +1,13 @@
 import { combineReducers, createReducer } from '@reduxjs/toolkit';
 
 import { type DeviceReducerState, prepareDeviceReducer } from '@suite-common/device';
+import { type FeesState } from '@suite-common/fee-types';
 import { createThunk } from '@suite-common/redux-utils';
 import { type TrezorDevice } from '@suite-common/suite-types';
 import { mockSuiteDevice } from '@suite-common/suite-types/mocks';
 import { configureMockStore, extraDependenciesCommonMock } from '@suite-common/test-utils';
 import { composeSendFormTransactionFeeLevelsThunk } from '@suite-common/wallet-core';
-import { type Account, type FeesState } from '@suite-common/wallet-types';
+import { type Account } from '@suite-common/wallet-types';
 import { type TokenInfo } from '@trezor/connect';
 
 import { accountBtc } from '../../../__fixtures__/utils';
