+++ suite-common/wallet-core/package.json
@@ -20,6 +20,7 @@
         "@suite-common/device": "workspace:*",
         "@suite-common/earn-stablecoin-api": "workspace:*",
         "@suite-common/earn-staking-api": "workspace:*",
+        "@suite-common/fee-types": "workspace:*",
         "@suite-common/fiat-services": "workspace:*",
         "@suite-common/firmware": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
