+++ suite-common/wallet-core/src/fees/feesActions.ts
@@ -1,6 +1,6 @@
 import { createAction } from '@reduxjs/toolkit';
 
-import { type FeesState } from '@suite-common/wallet-types';
+import { type FeesState } from '@suite-common/fee-types';
 
 export const FEES_MODULE_PREFIX = '@common/wallet-core/fees';
