+++ suite-common/wallet-core/tsconfig.json
@@ -8,6 +8,7 @@
         { "path": "../device" },
         { "path": "../earn-stablecoin-api" },
         { "path": "../earn-staking-api" },
+        { "path": "../fee-types" },
         { "path": "../fiat-services" },
         { "path": "../firmware" },
         { "path": "../redux-utils" },
