+++ suite-common/wallet-types/src/transaction.ts
@@ -92,26 +92,6 @@ export type Output = {
     dataAscii?: string; // bitcoin opreturn/ethereum data
 };
 
-export interface FeeInfo {
-    blockHeight: number; // when fee info was updated; 0 = never
-    blockTime: number; // how often block is mined
-    minFee: number;
-    maxFee: number;
-    minPriorityFee: number; // eth minimum max priority fee
-    dustLimit?: number; // coin dust limit
-    feeLimit?: number; // eth gas limit
-    levels: FeeLevel[]; // fee levels are predefined in @trezor/connect > trezor-firmware/common
-}
-
-export type FeesStatus = 'preloaded' | 'loading' | 'loaded' | 'error';
-
-export type FeesState = {
-    [key in NetworkSymbol]?: {
-        status: FeesStatus;
-        data?: FeeInfo;
-    };
-};
-
 export type EthTransactionData = {
     token?: TokenInfo;
     chainId: number;
