+++ suite-common/wallet-types/tsconfig.json
@@ -2,6 +2,7 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
+        { "path": "../fee-types" },
         { "path": "../metadata-types" },
         { "path": "../wallet-config" },
         { "path": "../wallet-constants" },
