+++ suite-common/wallet-utils/package.json
@@ -17,6 +17,7 @@
         "@suite-common/address": "workspace:*",
         "@suite-common/calldata": "workspace:*",
         "@suite-common/earn-staking-api": "workspace:*",
+        "@suite-common/fee-types": "workspace:*",
         "@suite-common/fiat-services": "workspace:*",
         "@suite-common/suite-constants": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
