+++ suite-native/module-earn/package.json
@@ -23,6 +23,7 @@
         "@suite-common/earn-stablecoin": "workspace:*",
         "@suite-common/earn-stablecoin-api": "workspace:*",
         "@suite-common/earn-staking-api": "workspace:*",
+        "@suite-common/fee-types": "workspace:*",
         "@suite-common/formatters": "workspace:*",
         "@suite-common/message-system": "workspace:*",
         "@suite-common/mev": "workspace:*",
