+++ suite-native/module-send/package.json
@@ -20,6 +20,7 @@
         "@suite-common/address": "workspace:*",
         "@suite-common/dependency-injection": "workspace:*",
         "@suite-common/device": "workspace:*",
+        "@suite-common/fee-types": "workspace:*",
         "@suite-common/formatters": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
         "@suite-common/suite-utils": "workspace:*",
