+++ suite-native/module-trading/tsconfig.json
@@ -7,6 +7,9 @@
             "path": "../../suite-common/dependency-injection"
         },
         { "path": "../../suite-common/device" },
+        {
+            "path": "../../suite-common/fee-types"
+        },
         { "path": "../../suite-common/flags" },
         {
             "path": "../../suite-common/formatters"
