+++ suite-native/transaction-management/package.json
@@ -19,6 +19,7 @@
         "@reduxjs/toolkit": "2.11.2",
         "@suite-common/dependency-injection": "workspace:*",
         "@suite-common/device": "workspace:*",
+        "@suite-common/fee-types": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
         "@suite-common/suite-utils": "workspace:*",
