+++ suite-native/transaction-management/tsconfig.json
@@ -7,6 +7,9 @@
         },
         { "path": "../../suite-common/device" },
         {
+            "path": "../../suite-common/fee-types"
+        },
+        {
             "path": "../../suite-common/redux-utils"
         },
         {
